Refactor: 로그인 및 JWT 토큰 발급 때 속도 개선을 위해 UserLoginHistory 저장은 비동기 처리로 리팩토링해라 - #103
