+++ tictoc-api/src/main/java/tictoc/user/service/UserCommandService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import tictoc.error.ErrorCode;
 import tictoc.error.exception.BadRequestException;
+import tictoc.kafka.evnt.publisher.LoginHistoryPublisher;
 import tictoc.kakao.KakaoFeignProvider;
 import tictoc.config.security.jwt.dto.JwtResDTO;
 import tictoc.config.security.jwt.util.JwtProvider;
@@ -23,6 +24,7 @@ import tictoc.user.port.UserRepositoryPort;
 public class UserCommandService implements UserCommandUseCase {
     private final UserRepositoryPort userRepositoryPort;
     private final ProfileRepositoryPort profileRepositoryPort;
+    private final LoginHistoryPublisher loginHistoryPublisher;
     private final JwtProvider jwtProvider;
     private final KakaoFeignProvider kakaoFeignProvider;
 
@@ -31,11 +33,11 @@ public class UserCommandService implements UserCommandUseCase {
         try {
             final var accessToken = kakaoFeignProvider.getKakaoAccessToken(authenticationCode);
             final var kakaoId = kakaoFeignProvider.getSocialId(accessToken);
-            if (userRepositoryPort.existsUserByKakaoId(kakaoId)) {
-                return createJwt(userRepositoryPort.findUserByKakaoId(kakaoId).getId());
-            } else {
-                return createUser(kakaoId, accessToken);
-            }
+            final Long userId = userRepositoryPort.findUserByKakaoId(kakaoId)
+                    .map(User::getId)
+                    .orElseGet(() -> createUser(kakaoId, accessToken));
+            loginHistoryPublisher.publish(userId);
+            return createJwt(userId);
         } catch (FeignException e) {
             throw new BadRequestException(ErrorCode.KAKAO_BAD_REQUEST);
         }
@@ -43,13 +45,13 @@ public class UserCommandService implements UserCommandUseCase {
 
     //TODO 카카오 소셜 로그아웃 기능 구현해야 함.
 
-    private JwtResDTO.Login createUser(String kakaoId, String accessToken) {
+    private Long createUser(String kakaoId, String accessToken) {
         KakaoResDTO.KakaoUserInfo kakaoUserInfo = kakaoFeignProvider.getKakaoProfile(accessToken);
 
         User user = saveUser(kakaoId, kakaoUserInfo.kakaoAccount().profile().nickname());
         Profile profile = saveProfile(user.getId(), kakaoUserInfo.kakaoAccount().profile().nickname());
         saveProfileImage(profile.getId(), kakaoUserInfo.kakaoAccount().profile().profileImgUrl());
-        return createJwt(user.getId());
+        return user.getId();
     }
 
     private JwtResDTO.Login createJwt(final Long userId) {
