+++ tictoc-domain/src/main/java/tictoc/user/adaptor/UserRepositoryAdaptor.java
@@ -6,6 +6,7 @@ import tictoc.user.exception.UserNotFoundException;
 import tictoc.user.model.User;
 import tictoc.user.port.UserRepositoryPort;
 import tictoc.user.repository.UserRepository;
+import java.util.Optional;
 import static tictoc.error.ErrorCode.USER_NOT_FOUND;
 
 @Component
@@ -17,12 +18,12 @@ public class UserRepositoryAdaptor implements UserRepositoryPort {
         return userRepository.save(user);
     }
 
-    public User findUserById(Long userId) {
-        return userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(USER_NOT_FOUND));
+    public Optional<User> findUserByKakaoId(String kakaoId) {
+        return userRepository.findByKakaoId(kakaoId);
     }
 
-    public User findUserByKakaoId(String kakaoId) {
-        return userRepository.findByKakaoId(kakaoId).orElseThrow(() -> new UserNotFoundException(USER_NOT_FOUND));
+    public User findUserById(Long userId) {
+        return userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(USER_NOT_FOUND));
     }
 
     public boolean existsUserByKakaoId(String kakaoId) {
